Allow animation to work when going North by fixing eigenvector calc.
My previous eigenvector based on an iterative algorithm was not
converging when attempting to find the eigenvector of the cumulative
rotation matrix for a square adjacent to the North.  So now I'm
using Jama to do that.

This change also adjusts base star drawing to filter for stars in
common with the active square.